+++ app.verify/D.MTExtDefinitions/01.IntFactorization/IF_NDTM.06GenNumber2.cs
@@ -151,7 +151,7 @@ namespace MTExtDefinitions
               {
                 new StateSymbolDirectionTriple()
                   {
-                    State = rejectingState,
+                    State = (int)rejectingState,
                     Symbol = delimiter,
                     Direction = TMDirection.S
                   }
