+++ app.verify/D.MTExtDefinitions/01.IntFactorization/IF_NDTM.07Multiply.cs
@@ -26,7 +26,7 @@ namespace MTExtDefinitions
               {
                 new StateSymbolDirectionTriple
                   {
-                    State = acceptingState,
+                    State = (int)acceptingState,
                     Symbol = delimiter,
                     Direction = TMDirection.S
                   }
